Fix usages of syscall constants for Windows
AIUI, the default underlying type of a C enum in MSVC/Clang on Windows is
`int`, regardless of whether there is a negative discriminant. Meanwhile,
on other platforms Clang/GCC use `unsigned int` if it can.

This causes bindgen to use different types for the generated module constants,
preventing it from compiling on Windows due to u32/i32 type mismatches[1].
While ugly, simply fixing it by casting to the appropriate type based on
context for now...

[1] https://github.com/rust-lang/rust-bindgen/issues/1907